chore: 단일 candy DTO 적용
